# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tchat API - Telegram SEA Edition
  description: |
    Comprehensive API documentation for Tchat (Telegram SEA Edition), a Southeast Asian messaging and commerce platform.

    ## Features
    - **Authentication**: OTP-based authentication with JWT tokens
    - **Messaging**: Real-time messaging with WebSocket support
    - **Payments**: Multi-currency wallet system with regional support
    - **Commerce**: E-commerce platform with shop and product management
    - **Notifications**: Multi-channel notification system with templates

    ## Regional Support
    - **Countries**: Thailand, Singapore, Indonesia, Malaysia, Philippines, Vietnam
    - **Currencies**: THB, SGD, IDR, MYR, PHP, VND, USD
    - **Languages**: English, Thai, Bahasa Indonesia, Bahasa Malaysia, Filipino, Vietnamese

    ## Security
    - JWT-based authentication with refresh tokens
    - Rate limiting on all endpoints
    - CORS support for web applications
    - Webhook signature verification

    ## WebSocket Support
    Real-time messaging is available via WebSocket connections at `/ws/messaging`
  version: "1.0.0"
  contact:
    name: Tchat API Support
    url: https://tchat.sea/support
    email: api-support@tchat.sea
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.tchat.sea/v1
    description: Production server
  - url: https://staging-api.tchat.sea/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from authentication endpoints
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request payload"
        details:
          type: string
          description: Additional error details
          example: "Field 'email' is required"
        code:
          type: string
          description: Error code for programmatic handling
          example: "VALIDATION_ERROR"
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items per page
          example: 20
        total:
          type: integer
          minimum: 0
          description: Total number of items
          example: 150
        total_pages:
          type: integer
          minimum: 0
          description: Total number of pages
          example: 8
        has_next:
          type: boolean
          description: Whether there is a next page
          example: true
        has_prev:
          type: boolean
          description: Whether there is a previous page
          example: false

    # Authentication schemas
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        recipient_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [email, sms, push, in_app]
          example: "push"
        channel:
          type: string
          example: "order_updates"
        subject:
          type: string
          example: "Your order has been shipped"
        content:
          type: string
          example: "Your order #12345 has been shipped and will arrive in 2-3 business days"
        status:
          type: string
          enum: [pending, sent, delivered, failed, expired]
          example: "delivered"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          example: "medium"
        read:
          type: boolean
          example: false
        read_at:
          type: string
          format: date-time
          nullable: true
        sent_at:
          type: string
          format: date-time
          nullable: true
        delivered_at:
          type: string
          format: date-time
          nullable: true
        failed_at:
          type: string
          format: date-time
          nullable: true
        scheduled_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        retry_count:
          type: integer
          example: 0
        metadata:
          type: object
          description: Additional notification data
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
            code: "UNAUTHORIZED"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            details: "Field 'email' is required"
            code: "VALIDATION_ERROR"
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded"
            code: "RATE_LIMIT_EXCEEDED"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
paths:
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get user notifications
      description: Retrieve notifications for the authenticated user with pagination
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: type
          in: query
          description: Filter by notification type
          schema:
            type: string
            enum: [email, sms, push, in_app]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [pending, sent, delivered, failed, expired]
        - name: unread_only
          in: query
          description: Show only unread notifications
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Notifications
      summary: Send notification
      description: Send a notification to a specific user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipient_id
                - type
                - channel
                - content
                - priority
              properties:
                recipient_id:
                  type: string
                  format: uuid
                type:
                  type: string
                  enum: [email, sms, push, in_app]
                channel:
                  type: string
                  example: "order_updates"
                template_id:
                  type: string
                  format: uuid
                subject:
                  type: string
                  example: "Your order update"
                content:
                  type: string
                  example: "Your order has been processed"
                variables:
                  type: object
                  description: Template variables
                priority:
                  type: string
                  enum: [low, medium, high, urgent]
                scheduled_at:
                  type: string
                  format: date-time
                expires_at:
                  type: string
                  format: date-time
                metadata:
                  type: object
      responses:
        '201':
          description: Notification sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification:
                    $ref: '#/components/schemas/Notification'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...